UI: one file per section panel
Structure-only commit, same recipe as the chart split: sections.rs
becomes a facade (mod declarations + re-exports, callers keep
addressing sections::*) and each of the ten panels lives in its own
sections/*.rs file, so a PR touching one panel names it in the
changed-file list. Private helpers stay with their only consumer
(cost_row / window_cost_usd in cost.rs, compact_duration in
parallel.rs); entry points are pub(in crate::ui) behind pub(super)
re-exports.

Also restores four doc-comment lines the chart-split extraction
dropped (charts/model_daily.rs, charts/limits.rs, and the same defect
in sections/cost.rs, sections/projects.rs) — verified by mechanical
comparison of every moved doc/attr block against the pre-split
originals. Second-model review ran two rounds to a clean verdict;
same-instant A/B render is identical apart from the load-time footer.
