+++ src/ui/charts/model_daily.rs
@@ -8,6 +8,7 @@ use crate::model::Summary;
 use super::{Y_AXIS_WIDTH, axis_label_row};
 use crate::ui::{theme, utils};
 
+/// Daily volume as stacked per-model bars, rendered by hand: one column per
 /// day (or per day-bucket on narrow terminals), each half-cell colored by
 /// the segment that owns it. The Chart widget painter-stacking left rounding
 /// artifacts (floating caps, bleeding columns); exact half-cell assignment
